fix: bump mocha version
Issue was actually with `serialize-javascript` which is a dependency of
`mocha`.

Related Advisory: https://github.com/advisories/GHSA-hxcc-f52p-wc94

Also changed the supported versions to the Node.js versions that are
still active and not at EOL and modernized/cleaned up the documentation
a bit.

Bumped the major version due to the major update to what versions of
Node.js we're testing against.